+++ components/auth/confirm-signup-form.tsx
@@ -11,7 +11,7 @@ import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
 import { Mail } from "lucide-react";
 import { ConfirmState, handleConfirmSignUp } from "@/lib/actions/signup-action";
-import { ErrorMessage } from "@/app/ui/error-message";
+import { ErrorMessage } from "@/components/forms/error-message";
 import { useActionState } from "react";
 
 export function ConfirmSignUpForm() {
