+++ components/auth/login-form.tsx
@@ -11,7 +11,7 @@ import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
 import Link from "next/link";
 import { useState, useActionState } from "react";
-import { ErrorMessage } from "@/app/ui/error-message";
+import { ErrorMessage } from "@/components/forms/error-message";
 import { handleSignIn, LoginState } from "@/lib/actions/signin-action";
 
 export function LoginForm() {
